fix issues: - use lowercase for folders - remove empty lines - rename redundant namings - rename unlogic naming
