+++ src/views/Events.vue
@@ -6,7 +6,7 @@
       </div>
       <div class="col-9 py-3">
         <h1>{{ headline }}</h1>
-        <app-overview></app-overview>
+        <app-event-listing></app-event-listing>
       </div>
     </div>
   </div>
@@ -14,12 +14,12 @@
 
 <script>
 import AppNavigation from './../components/Navigation'
-import AppOverview from './../components/Events/Overview'
+import AppEventListing from '../components/events/Listing'
 
 export default {
   components: {
     AppNavigation,
-    AppOverview
+    AppEventListing
   },
   data () {
     return {
